refactor(apps,guard): four dense functions become the shapes they describe
`vendo_make`, `validatePlan`, the Claude session loop and the guard's bound
`execute` each carried one long function doing several jobs. Every one of them
is decomposed here, and only decomposed: no bug fixed, no name changed beyond
what a move requires, no test touched.

- apps/agent-tools.ts — `execute` is a dispatcher. `vendo_make`'s two routes
  (assemble-or-escalate, change-a-named-app) move to `make-tool.ts`, the three
  `vendo_apps_data_*` doors to `data-tools.ts`, and the argument checks all of
  them share to `tool-args.ts`.
- apps/automation-plan.ts — `validatePlan` hands its `steps` rules to a
  `stepsIssues` collector beside the `scheduleIssues` one it already had.
- apps/claude-turn.ts — `sessionOptions` and `readAssistantMessage` are named
  siblings IN THIS FILE, not modules: `dist/claude-turn.js` is copied verbatim
  into the box image, so it may hold no relative import.
- guard/guard.ts — `bind().execute` keeps the decisions and hands the dispatch
  to `#runOnce`: the grant, the effect key, the in-flight share and the receipt
  write now sit together.

No public surface changed. Zero test files changed.
